examples/flask_app/requirements_for_examples.txt
blinker==1.4
cachelib==0.2.0
click==8.0.1
Flask==2.0.1
Flask-DebugToolbar==0.11.0
Flask-Session==0.4.0
Flask-WTF==0.15.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Pillow==8.2.0
python-dotenv==0.18.0
swatcher==1.0.0
Werkzeug==2.0.1
WTForms==2.3.3
